Add user events directory and message module
